requirements.txt
alabaster==0.7.12
allure-pytest==2.9.45
allure-python-commons==2.9.45
app==0.0.1
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
Babel==2.10.3
certifi==2022.5.18.1
cffi
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==37.0.2
docutils==0.18.1
h11==0.13.0
idna==3.3
imagesize==1.4.1
importlib-metadata==4.12.0
iniconfig==1.1.1
Jinja2==3.1.2
jsonpickle==2.2.0
MarkupSafe==2.1.1
mimesis==5.4.0
outcome==1.1.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
pytest-rerunfailures==10.2
pytz==2022.1
requests==2.28.0
selenium==4.2.0
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
Sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.9
wsproto==1.1.0
zipp==3.8.0